Add Twitch-Discord Listener Cog
Adds behavior to Discord bot that checks a specific Twitch livestream once every 5 seconds. When the twitch stream becomes live, it sends a message into a channel on Discord. The Twitch and Discord bots are both currently static, and must be configured in the .env file.

Coexists with the previous code, but bot.py needs to be updated to properly add this cog.